patch: Moved DataBaseConnection class to Helpers\DB
