+++ Idler/ShiftNote.cs
@@ -1,4 +1,5 @@
-﻿using Idler.Interfaces;
+﻿using Idler.Helpers.DB;
+using Idler.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Data;
